Server config

+++ code/emilie-h_Mystere98/Server/src/main/java/ch/heig/dai/lab/protocoldesign/operation/Mul.java
@@ -1,7 +1,7 @@
 package ch.heig.dai.lab.protocoldesign.operation;
 
 public class Mul {
-    public static int make(int a, int b) {
+    public static double make(int a, int b) {
         return a * b;
     }
 }
